доработка фильтрации
