+++ index.html
@@ -237,11 +237,10 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script async src="js/page.js" type="module"></script>
-  <script async src="js/main.js" type="module"></script>
-  <script async src="pristine/pristine.min.js"></script>
-  <script async src="nouislider/nouislider.js"></script>
-  <script async src="leaflet/leaflet.js"></script>
-  <script async src="leaflet/leaflet.js.map"></script>
+  <script src="js/page.js" type="module"></script>
+  <script src="pristine/pristine.min.js"></script>
+  <script src="nouislider/nouislider.js"></script>
+  <script src="leaflet/leaflet.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
