+++ js/filter.js
@@ -1,72 +1,64 @@
-import {mapFilter} from './page.js';
-import {initialAds, OFFERS_COUNT, createMarker} from './map.js';
+import {initialAds, OFFERS_COUNT, createMarker, markerGroup} from './map.js';
+import {debounce} from './util.js';
 
+const FILTER_DELAY = 500;
 
-const DEFAULT_FILTER_VALUE = 'any';
+const filterElement = document.querySelector('.map__filters-container');
 const filterTypeElement = document.querySelector('[name="housing-type"]');
-let filterTypeValue = DEFAULT_FILTER_VALUE;
 const filterPriceElement = document.querySelector('[name="housing-price"]');
-let filterPriceValue = DEFAULT_FILTER_VALUE;
 const filterRoomsElement = document.querySelector('[name="housing-rooms"]');
-let filterRoomsValue = DEFAULT_FILTER_VALUE;
 const filterGuestsElement = document.querySelector('[name="housing-guests"]');
-let filterGuestsValue = DEFAULT_FILTER_VALUE;
 const featureElements = Array.from(document.querySelectorAll('[name="features"]'));
 
 
 //фильтрация по типу жилья
 const filterType = (ad) => {
-  filterTypeValue = filterTypeElement.value;
+  const filterTypeValue = filterTypeElement.value;
   if (filterTypeValue === 'any') {
     return true;
-  } else {
-    return filterTypeValue === ad.offer.type;
   }
+  return filterTypeValue === ad.offer.type;
 };
 
 //фильтрация по цене
-const filterPrices = {
+const FilterPrices = {
   low: {
-    min: 0,
-    max: 10000,
+    MIN: 0,
+    MAX: 10000,
   },
   middle: {
-    min: 10000,
-    max: 50000,
+    MIN: 10000,
+    MAX: 50000,
   },
   high: {
-    min: 50000,
-    max: 100000,
+    MIN: 50000,
+    MAX: 100000,
   },
 };
 const filterPrice = (ad) => {
-  filterPriceValue = filterPriceElement.value;
+  const filterPriceValue = filterPriceElement.value;
   if (filterPriceValue === 'any') {
     return true;
-  } else {
-
-    return ad.offer.price > filterPrices[filterPriceValue].min && ad.offer.price < filterPrices[filterPriceValue].max;
   }
+  return ad.offer.price > FilterPrices[filterPriceValue].MIN && ad.offer.price < FilterPrices[filterPriceValue].MAX;
 };
 
 //филтрация по числу комнат
 const filterRooms = (ad) => {
-  filterRoomsValue = filterRoomsElement.value;
+  const filterRoomsValue = filterRoomsElement.value;
   if (filterRoomsValue === 'any') {
     return true;
-  } else {
-    return filterRoomsValue === String(ad.offer.rooms);
   }
+  return filterRoomsValue === String(ad.offer.rooms);
 };
 
 //фильтрация по числу гостей
 const filterGuests = (ad) => {
-  filterGuestsValue = filterGuestsElement.value;
+  const filterGuestsValue = filterGuestsElement.value;
   if (filterGuestsValue === 'any') {
     return true;
-  } else {
-    return filterGuestsValue === String(ad.offer.guests);
   }
+  return filterGuestsValue === String(ad.offer.guests);
 };
 
 //фильтрация фичеры
@@ -76,12 +68,15 @@ const filterFeatures = (ad) => {
 };
 
 //сброс маркеров на карте
-const mapFilterReset = () => {
-  mapFilter.reset();
-  initialAds.slice(0, OFFERS_COUNT)
+
+
+filterElement.addEventListener('change', debounce(() => {
+  markerGroup.clearLayers();
+  initialAds
+    .filter((ad) => filterGuests(ad) && filterRooms(ad) && filterPrice(ad) && filterType(ad) && filterFeatures(ad))
+    .slice(0, OFFERS_COUNT)
     .forEach((ad) => {
       createMarker(ad);
     });
-};
+}, FILTER_DELAY));
 
-export {filterGuests, filterRooms, filterPrice, filterType, filterFeatures, mapFilterReset};
