+++ js/main.js
@@ -3,4 +3,4 @@ import './form.js';
 import './add-card.js';
 import './map.js';
 import './photo.js';
-
+import './filter.js';
